feat(engine): Tensor Core GEMMs, FlashAttention, Fused Kernels, Static Arena, Multi-GPU P2P, and Docs
- cuBLAS Tensor Core GEMMs with FP32/TF32, FP16, and BF16 precision support
- Tiled FlashAttention forward and backward kernels with O(N) memory complexity
- Fused CUDA kernels: LayerNorm warp-shuffle reduction, fused AdamW, bias + activations
- CUDA Graphs capture and replay infrastructure
- CUDAScratchpadArena zero-allocation memory arena with RAII ArenaScope
- Native multi-GPU pipeline with DeviceMesh and peer-to-peer NVLink/PCIe ring reduction
- Complete IEEE 754 float16 and bfloat16 bitwise precision emulation and conversions
- Modernized README with architectural diagrams, benchmarks, and pure C++ training examples

+++ src/cool_ops.h
@@ -29,7 +29,7 @@ struct ScaleAndShiftBackward : public Node {
         float* gbe_ptr = gbe->data_ptr();
 
         for (int batch = 0; batch < batch_size; batch++) {
-            std::vector<int> batch_idx = unravel(batch, std::vector<int>(x->shape.begin(), x->shape.end() - 2));
+            std::vector<int64_t> batch_idx = unravel(batch, std::vector<int64_t>(x->shape.begin(), x->shape.end() - 2));
 
             int batch_off_x = 0;
             int batch_off_out = 0;
@@ -78,7 +78,7 @@ inline std::shared_ptr<Tensor> scale_n_shift(const std::shared_ptr<Tensor>& x,
     float* out_ptr = out->data_ptr();
 
     for (int batch = 0; batch < batch_size; batch++) {
-        std::vector<int> batch_idx = unravel(batch, std::vector<int>(x->shape.begin(), x->shape.end() - 2));
+        std::vector<int64_t> batch_idx = unravel(batch, std::vector<int64_t>(x->shape.begin(), x->shape.end() - 2));
 
         int batch_off_x = 0;
         int batch_off_out = 0;
